Creación de modulos Grupo, Aulas y Cursos

+++ escuela/src/main/java/com/gina/escuela/services/aulas/AulaServiceImpl.java
@@ -2,11 +2,14 @@ package com.gina.escuela.services.aulas;
 
 import com.gina.escuela.dto.aula.AulaRequest;
 import com.gina.escuela.dto.aula.AulaResponse;
+import com.gina.escuela.dto.maestro.MaestroRequest;
 import com.gina.escuela.entities.Aula;
 import com.gina.escuela.entities.Maestro;
+import com.gina.escuela.exceptions.EntidadRelacionadaException;
 import com.gina.escuela.exceptions.RecursoNoEncontradoException;
 import com.gina.escuela.mappers.AulaMapper;
 import com.gina.escuela.repositories.AulaRepository;
+import com.gina.escuela.repositories.GrupoRepository;
 import com.gina.escuela.utils.ServiceUtils;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,6 +26,8 @@ public class AulaServiceImpl implements AulaService{
 
     private final AulaRepository aulaRepository;
     private final AulaMapper aulaMapper;
+
+    private final GrupoRepository grupoRepository;
 
     @Override
     @Transactional(readOnly = true)
@@ -43,6 +48,7 @@ public class AulaServiceImpl implements AulaService{
     @Override
     public AulaResponse registrar(AulaRequest request) {
         log.info("Registrando aula...");
+        validarNombreUnico(request);
         Aula aula = aulaMapper.requestAEntidad(request);
         aulaRepository.save(aula);
         log.info("Nueva aula {} registrada", aula.getNombre());
@@ -51,8 +57,9 @@ public class AulaServiceImpl implements AulaService{
 
     @Override
     public AulaResponse actualizar(AulaRequest request, Long id) {
-        Aula aula = obtenerAulaExcepcion(id);
         log.info("Actualizando aula con id: {}", id);
+        Aula aula = obtenerAulaExcepcion(id);
+        validarNombreUnicoActualizar(request, id);
         aula.actualizar(
                 request.nombre(),
                 request.capacidad()
@@ -64,12 +71,28 @@ public class AulaServiceImpl implements AulaService{
 
     @Override
     public void eliminar(Long id) {
+        log.info("Eliminando aula con id: {}", id);
         Aula aula = obtenerAulaExcepcion(id);
+
+        if (grupoRepository.existsByAulaId(id))
+            throw new EntidadRelacionadaException("No se puede eliminar al aula ya que tiene grupos asignados");
         aulaRepository.delete(aula);
         log.info("Aula con id: {} eliminada", id);
     }
 
     private Aula obtenerAulaExcepcion(Long id){
         return ServiceUtils.obtenerEntidadOException(aulaRepository, id, Aula.class);
+    }
+
+    private void validarNombreUnico(AulaRequest request){
+        log.info("Validando nombre de aula único...");
+        if (aulaRepository.existsByNombreIgnoreCase(request.nombre()))
+            throw  new IllegalArgumentException("Ya existe un aula registrada con el nombre: " + request.nombre());
+   }
+
+    private void validarNombreUnicoActualizar(AulaRequest request, Long id){
+        log.info("Validando cambio de nombre de aula único...");
+        if (aulaRepository.existsByNombreIgnoreCaseAndIdNot(request.nombre(), id))
+            throw  new IllegalArgumentException("Ya existe un aula registrada con el nombre: " + request.nombre());
     }
 }
