+++ escuela/src/main/java/com/gina/escuela/services/cursos/CursoServiceImpl.java
@@ -1,12 +1,15 @@
 package com.gina.escuela.services.cursos;
 
+import com.gina.escuela.dto.aula.AulaRequest;
 import com.gina.escuela.dto.curso.CursoRequest;
 import com.gina.escuela.dto.curso.CursoResponse;
 import com.gina.escuela.entities.Curso;
 import com.gina.escuela.entities.Maestro;
+import com.gina.escuela.exceptions.EntidadRelacionadaException;
 import com.gina.escuela.exceptions.RecursoNoEncontradoException;
 import com.gina.escuela.mappers.CursoMapper;
 import com.gina.escuela.repositories.CursoRepository;
+import com.gina.escuela.repositories.GrupoRepository;
 import com.gina.escuela.utils.ServiceUtils;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,6 +26,7 @@ public class CursoServiceImpl implements CursoService {
 
     private final CursoRepository cursoRepository;
     private final CursoMapper cursoMapper;
+    private final GrupoRepository grupoRepository;
 
     @Override
     @Transactional(readOnly = true)
@@ -43,6 +47,7 @@ public class CursoServiceImpl implements CursoService {
     @Override
     public CursoResponse registrar(CursoRequest request) {
         log.info("Registrando curso...");
+        validarNombreUnico(request);
         Curso curso = cursoMapper.requestAEntidad(request);
         cursoRepository.save(curso);
         log.info("Nuevo curso {} registrado", curso.getNombre());
@@ -51,8 +56,9 @@ public class CursoServiceImpl implements CursoService {
 
     @Override
     public CursoResponse actualizar(CursoRequest request, Long id) {
-        Curso curso = obtenerCursoExcepcion(id);
         log.info("Actualizando curso con id: {}", id);
+        validarNombreUnicoActualizar(request, id);
+        Curso curso = obtenerCursoExcepcion(id);
         curso.actualizar(
                 request.nombre(),
                 request.descripcion(),
@@ -65,12 +71,28 @@ public class CursoServiceImpl implements CursoService {
 
     @Override
     public void eliminar(Long id) {
+        log.info("Eliminando curso con id: {}", id);
         Curso curso = obtenerCursoExcepcion(id);
+        if (grupoRepository.existsByCursoId(id))
+            throw new EntidadRelacionadaException("No se puede eliminar al curso ya que tiene grupos asignados");
+
         cursoRepository.delete(curso);
         log.info("Curso con id: {} eliminado", id);
     }
 
     private Curso obtenerCursoExcepcion(Long id){
         return ServiceUtils.obtenerEntidadOException(cursoRepository, id, Curso.class);
+    }
+
+    private void validarNombreUnico(CursoRequest request){
+        log.info("Validando nombre de curso único...");
+        if (cursoRepository.existsByNombreIgnoreCase(request.nombre()))
+            throw  new IllegalArgumentException("Ya existe un curso registrado con el nombre: " + request.nombre());
+    }
+
+    private void validarNombreUnicoActualizar(CursoRequest request, Long id){
+        log.info("Validando cambio de nombre de curso único...");
+        if (cursoRepository.existsByNombreIgnoreCaseAndIdNot(request.nombre(), id))
+            throw  new IllegalArgumentException("Ya existe un curso registrado con el nombre: " + request.nombre());
     }
 }
